Long Data support Added support to files whose number of bytes is larger than 2^32

+++ cell.h
@@ -1,6 +1,6 @@
 typedef struct c_cell{
 	char key[30];
-	int value;
+	long int value;
 } cell;
 
 cell* new_cell();
